# contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: Structura API
  version: 0.1.3
  summary: Local-first AI-augmented document filing cabinet API
  description: 'This v1.3 API contract covers the core workbench surfaces needed
    for implementation convergence: auth/session, document ingest and detail,
    protected asset access, folders, tags, candidate/canonical review actions,
    relationships, contacts, filing rules, watched folders, search, analysis,
    exports, and job/admin inspection.

    '
servers:
- url: http://localhost:8000
  description: Local development
...
security:
- sessionCookie: []
- apiToken: []
paths:
...
  /api/v1/migrations/baseline:
    get:
      tags:
      - Admin
      summary: Baseline migration manifest
      security:
      - sessionCookie: []
      - apiToken: []
      responses:
        '200':
          description: Baseline migration manifest
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - database_dir
                - scripts
                properties:
                  database_dir:
                    type: string
                  scripts:
                    type: array
                    items:
                      type: string
        '401':
          description: Not authenticated
  /api/v1/auth/session:
    get:
      tags:
      - Auth
      summary: Get current session
      security:
      - sessionCookie: []
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '401':
          description: Not authenticated
    post:
      tags:
      - Auth
      summary: Create a login session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Authenticated session created
          headers:
            Set-Cookie:
              description: HttpOnly session cookie carrying the new session token.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '401':
          description: Authentication failed
    delete:
      tags:
      - Auth
      summary: End current session
      security:
      - sessionCookie: []
      responses:
        '204':
          description: Session ended
          headers:
            Set-Cookie:
              description: Session cookie cleared by expiring the current cookie value.
              schema:
                type: string
        '401':
          description: Not authenticated
  /api/v1/auth/magic-links:
    post:
      tags:
      - Auth
      summary: Request a magic link for bootstrap, invite, or recovery
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMagicLinkRequest'
      responses:
        '202':
          description: Magic link requested
...
  /api/v1/jobs/{jobId}:
    get:
      tags:
      - Jobs
      summary: Fetch job status
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobState'
        '401':
          description: Not authenticated
  /api/v1/admin/jobs:
    get:
      tags:
      - Admin
      summary: List jobs for admin and dead-letter triage
      parameters:
      - in: query
        name: status
        schema:
          type: string
      - in: query
        name: jobType
        schema:
          type: string
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobState'
        '401':
          description: Not authenticated
  /api/v1/admin/jobs/{jobId}/retry:
    post:
      tags:
      - Admin
      summary: Retry a failed or dead-letter job
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: Retry accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJob'
        '401':
          description: Not authenticated
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: structura_session
      description: HttpOnly browser session cookie issued by `POST /api/v1/auth/session`.
    apiToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: Scoped automation token for CLI, watched-folder, and admin flows.
  schemas:
    AcceptedJob:
      type: object
      additionalProperties: false
      required:
      - jobId
      - status
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - queued
          - running
...
    JobState:
      type: object
      additionalProperties: false
      required:
      - jobId
      - jobType
      - status
      - createdAt
      properties:
        jobId:
          type: string
          format: uuid
        jobType:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        errorMessage:
          type: string
        result:
          type: object
          additionalProperties: true
...
    CreateSessionRequest:
      oneOf:
      - $ref: '#/components/schemas/PasswordSessionRequest'
      - $ref: '#/components/schemas/MagicLinkSessionRequest'
      discriminator:
        propertyName: method
        mapping:
          password: '#/components/schemas/PasswordSessionRequest'
          magic_link: '#/components/schemas/MagicLinkSessionRequest'
    PasswordSessionRequest:
      type: object
      additionalProperties: false
      required:
      - method
      - email
      - password
      properties:
        method:
          type: string
          const: password
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        householdId:
          type: string
          format: uuid
    MagicLinkSessionRequest:
      type: object
      additionalProperties: false
      required:
      - method
      - magicLinkToken
      properties:
        method:
          type: string
          const: magic_link
        magicLinkToken:
          type: string
          minLength: 1
        householdId:
          type: string
          format: uuid
    SessionInfo:
      type: object
      additionalProperties: false
      required:
      - sessionId
      - userId
      - isAuthenticated
      - authMethod
      properties:
        sessionId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        displayName:
          type: string
        email:
          type: string
          format: email
        authMethod:
          type: string
          enum:
          - password
          - magic_link
          - webauthn
          description: Authentication method persisted on the session record.
        isAuthenticated:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        passwordRotationRequired:
          type: boolean
          description: Present when a bootstrap password remains active and should be rotated.
    CreateMagicLinkRequest:
      type: object
      additionalProperties: false
      required:
      - email
      - purpose
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          enum:
          - bootstrap
          - invite
          - recovery
        householdId:
          type: string
          format: uuid